refactor(web): move withUploadServeScope out of the XHR uploader module
PromptComposer and ChatAttachments mock @/lib/api/uploads wholesale, so a
render helper exported from there came back undefined in their specs (5
failures in the full suite). It is a pure URL builder, not transport; it
gets its own module and spec.
