+++ apps/web/src/lib/api/uploads.unit.spec.ts
@@ -1,6 +1,6 @@
 // @vitest-environment jsdom
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
-import { UploadError, uploadFile, withUploadServeScope } from './uploads';
+import { UploadError, uploadFile } from './uploads';
 
 /**
  * `uploadFile` is the one browser→BFF call on a raw XMLHttpRequest (upload
@@ -115,48 +115,5 @@ describe('uploadFile — workspace selector on the XHR', () => {
 
         await expect(pending).rejects.toBeInstanceOf(UploadError);
         await expect(pending).rejects.toMatchObject({ message: 'File too large', status: 413 });
-    });
-});
-
-/**
- * The render-time half for the serve URL. The URL the API mints, the one in
- * chat text and the one in attachment lists stay scope-free; the tab's
- * workspace is added where an `<a href>` / `<img src>` is rendered.
- */
-describe('withUploadServeScope', () => {
-    const ORG = { kind: 'organization', slug: 'acme' } as const;
-
-    it('appends the tab’s selector to an API-minted serve URL', () => {
-        expect(withUploadServeScope('/api/uploads/u1/aaaa.png', ORG)).toBe(
-            '/api/uploads/u1/aaaa.png?scope=org%3Aacme',
-        );
-    });
-
-    it('keeps an existing workId next to the selector', () => {
-        const url = new URL(
-            withUploadServeScope('/api/uploads/u1/aaaa.png?workId=w-1', ORG),
-            'http://n',
-        );
-
-        expect(url.searchParams.get('workId')).toBe('w-1');
-        expect(url.searchParams.get('scope')).toBe('org:acme');
-    });
-
-    it('leaves the URL alone when there is no scope to add', () => {
-        expect(withUploadServeScope('/api/uploads/u1/aaaa.png', null)).toBe(
-            '/api/uploads/u1/aaaa.png',
-        );
-    });
-
-    it.each([
-        'blob:http://web.example/123',
-        'https://cdn.example/x.png',
-        '/api/works/w/kb/uploads/u/download',
-    ])('never decorates %j — only the serve URL family carries this', (url) => {
-        expect(withUploadServeScope(url, ORG)).toBe(url);
-    });
-
-    it('passes undefined through for an attachment with no URL yet', () => {
-        expect(withUploadServeScope(undefined, ORG)).toBeUndefined();
     });
 });
